+++ decode_test.go
@@ -83,7 +83,7 @@ func TestDecode(t *testing.T) {
 		},
 		{
 			dec: Tested{"test", []byte("test"), []string{"1", "2"}, false, 25, "", ""},
-			obj: map[interface{}]interface{}{"Name": "test", "data": []byte("test"), "Temp": []interface{}{"1", "2"}, "Count": int8(25)},
+			obj: map[interface{}]interface{}{"Name": "test", "data": []byte("test"), "Temp": []interface{}{"1", "2"}, "Count": int64(25)},
 		},
 		// ------------------------------------------------------------------------------------------------------------------------
 		// ------------------------------------------------------------------------------------------------------------------------
@@ -102,39 +102,39 @@ func TestDecode(t *testing.T) {
 		},
 		{
 			dec: []int8{1, math.MaxInt8},
-			obj: []interface{}{int8(1), int8(math.MaxInt8)},
+			obj: []interface{}{int64(1), int64(math.MaxInt8)},
 		},
 		{
 			dec: []int16{1, math.MaxInt8, math.MaxInt16},
-			obj: []interface{}{int8(1), int8(math.MaxInt8), int16(math.MaxInt16)},
+			obj: []interface{}{int64(1), int64(math.MaxInt8), int64(math.MaxInt16)},
 		},
 		{
 			dec: []int32{1, math.MaxInt8, math.MaxInt16, math.MaxInt32},
-			obj: []interface{}{int8(1), int8(math.MaxInt8), int16(math.MaxInt16), int32(math.MaxInt32)},
+			obj: []interface{}{int64(1), int64(math.MaxInt8), int64(math.MaxInt16), int64(math.MaxInt32)},
 		},
 		{
 			dec: []int64{1, math.MaxInt8, math.MaxInt16, math.MaxInt32, math.MaxInt64},
-			obj: []interface{}{int8(1), int8(math.MaxInt8), int16(math.MaxInt16), int32(math.MaxInt32), int64(math.MaxInt64)},
+			obj: []interface{}{int64(1), int64(math.MaxInt8), int64(math.MaxInt16), int64(math.MaxInt32), int64(math.MaxInt64)},
 		},
 		{
 			dec: []int{0, 1, math.MaxInt8, math.MaxInt16, math.MaxInt32, math.MaxInt64},
-			obj: []interface{}{int8(0), int8(1), int8(math.MaxInt8), int16(math.MaxInt16), int32(math.MaxInt32), int64(math.MaxInt64)},
+			obj: []interface{}{int64(0), int64(1), int64(math.MaxInt8), int64(math.MaxInt16), int64(math.MaxInt32), int64(math.MaxInt64)},
 		},
 		{
 			dec: []uint16{1, math.MaxUint8, math.MaxUint16},
-			obj: []interface{}{int8(1), uint8(math.MaxUint8), uint16(math.MaxUint16)},
+			obj: []interface{}{int64(1), uint64(math.MaxUint8), uint64(math.MaxUint16)},
 		},
 		{
 			dec: []uint32{1, math.MaxUint8, math.MaxUint16, math.MaxUint32},
-			obj: []interface{}{int8(1), uint8(math.MaxUint8), uint16(math.MaxUint16), uint32(math.MaxUint32)},
+			obj: []interface{}{int64(1), uint64(math.MaxUint8), uint64(math.MaxUint16), uint64(math.MaxUint32)},
 		},
 		{
 			dec: []uint64{1, math.MaxUint8, math.MaxUint16, math.MaxUint32, math.MaxUint64},
-			obj: []interface{}{int8(1), uint8(math.MaxUint8), uint16(math.MaxUint16), uint32(math.MaxUint32), uint64(math.MaxUint64)},
+			obj: []interface{}{int64(1), uint64(math.MaxUint8), uint64(math.MaxUint16), uint64(math.MaxUint32), uint64(math.MaxUint64)},
 		},
 		{
 			dec: []uint{0, 1, math.MaxUint8, math.MaxUint16, math.MaxUint32, math.MaxUint64},
-			obj: []interface{}{int8(0), int8(1), uint8(math.MaxUint8), uint16(math.MaxUint16), uint32(math.MaxUint32), uint64(math.MaxUint64)},
+			obj: []interface{}{int64(0), int64(1), uint64(math.MaxUint8), uint64(math.MaxUint16), uint64(math.MaxUint32), uint64(math.MaxUint64)},
 		},
 		{
 			dec: []float32{math.Pi, math.Pi},
@@ -153,12 +153,12 @@ func TestDecode(t *testing.T) {
 			obj: []interface{}{complex128(math.Pi), complex128(math.Pi)},
 		},
 		{
-			dec: []interface{}{nil, true, false, "test", []byte("test"), int8(77), uint8(177), float64(math.Pi)},
-			obj: []interface{}{nil, true, false, "test", []byte("test"), int8(77), uint8(177), float64(math.Pi)},
+			dec: []interface{}{nil, true, false, "test", []byte("test"), int64(77), uint64(177), float64(math.Pi)},
+			obj: []interface{}{nil, true, false, "test", []byte("test"), int64(77), uint64(177), float64(math.Pi)},
 		},
 		{
 			dec: []Tested{{"test", []byte("test"), []string{"1", "2"}, false, 25, "", ""}},
-			obj: []interface{}{map[interface{}]interface{}{"Name": "test", "data": []byte("test"), "Temp": []interface{}{"1", "2"}, "Count": int8(25)}},
+			obj: []interface{}{map[interface{}]interface{}{"Name": "test", "data": []byte("test"), "Temp": []interface{}{"1", "2"}, "Count": int64(25)}},
 		},
 		// ------------------------------------------------------------------------------------------------------------------------
 		// ------------------------------------------------------------------------------------------------------------------------
@@ -185,53 +185,53 @@ func TestDecode(t *testing.T) {
 		// ------------------------------------------------------------------------------------------------------------------------
 		{
 			dec: map[int]int{1: 1, 2: math.MaxInt64},
-			obj: map[interface{}]interface{}{int8(1): int8(1), int8(2): int64(math.MaxInt64)},
+			obj: map[interface{}]interface{}{int64(1): int64(1), int64(2): int64(math.MaxInt64)},
 		},
 		{
 			dec: map[int]uint{1: 1, 2: math.MaxUint64},
-			obj: map[interface{}]interface{}{int8(1): int8(1), int8(2): uint64(math.MaxUint64)},
+			obj: map[interface{}]interface{}{int64(1): int64(1), int64(2): uint64(math.MaxUint64)},
 		},
 		{
 			dec: map[int]bool{1: true, 2: false},
-			obj: map[interface{}]interface{}{int8(1): true, int8(2): false},
+			obj: map[interface{}]interface{}{int64(1): true, int64(2): false},
 		},
 		{
 			dec: map[int]string{1: "Hello", 2: "World"},
-			obj: map[interface{}]interface{}{int8(1): "Hello", int8(2): "World"},
+			obj: map[interface{}]interface{}{int64(1): "Hello", int64(2): "World"},
 		},
 		{
 			dec: map[int]interface{}{1: "Hello", 2: math.Pi},
-			obj: map[interface{}]interface{}{int8(1): "Hello", int8(2): float64(math.Pi)},
+			obj: map[interface{}]interface{}{int64(1): "Hello", int64(2): float64(math.Pi)},
 		},
 		// ------------------------------------------------------------------------------------------------------------------------
 		{
 			dec: map[uint]int{1: 1, 2: math.MaxInt64},
-			obj: map[interface{}]interface{}{int8(1): int8(1), int8(2): int64(math.MaxInt64)},
+			obj: map[interface{}]interface{}{int64(1): int64(1), int64(2): int64(math.MaxInt64)},
 		},
 		{
 			dec: map[uint]uint{1: 1, 2: math.MaxUint64},
-			obj: map[interface{}]interface{}{int8(1): int8(1), int8(2): uint64(math.MaxUint64)},
+			obj: map[interface{}]interface{}{int64(1): int64(1), int64(2): uint64(math.MaxUint64)},
 		},
 		{
 			dec: map[uint]bool{1: true, 2: false},
-			obj: map[interface{}]interface{}{int8(1): true, int8(2): false},
+			obj: map[interface{}]interface{}{int64(1): true, int64(2): false},
 		},
 		{
 			dec: map[uint]string{1: "Hello", 2: "World"},
-			obj: map[interface{}]interface{}{int8(1): "Hello", int8(2): "World"},
+			obj: map[interface{}]interface{}{int64(1): "Hello", int64(2): "World"},
 		},
 		{
 			dec: map[uint]interface{}{1: "Hello", 2: math.Pi},
-			obj: map[interface{}]interface{}{int8(1): "Hello", int8(2): float64(math.Pi)},
+			obj: map[interface{}]interface{}{int64(1): "Hello", int64(2): float64(math.Pi)},
 		},
 		// ------------------------------------------------------------------------------------------------------------------------
 		{
 			dec: map[string]int{"one": 1, "two": math.MaxInt64},
-			obj: map[interface{}]interface{}{"one": int8(1), "two": int64(math.MaxInt64)},
+			obj: map[interface{}]interface{}{"one": int64(1), "two": int64(math.MaxInt64)},
 		},
 		{
 			dec: map[string]uint{"one": 1, "two": math.MaxUint64},
-			obj: map[interface{}]interface{}{"one": int8(1), "two": uint64(math.MaxUint64)},
+			obj: map[interface{}]interface{}{"one": int64(1), "two": uint64(math.MaxUint64)},
 		},
 		{
 			dec: map[string]bool{"one": true, "two": false},
@@ -247,8 +247,8 @@ func TestDecode(t *testing.T) {
 		},
 		// ------------------------------------------------------------------------------------------------------------------------
 		{
-			dec: map[interface{}]interface{}{"one": "Hello", int8(2): float64(math.Pi)},
-			obj: map[interface{}]interface{}{"one": "Hello", int8(2): float64(math.Pi)},
+			dec: map[interface{}]interface{}{"one": "Hello", int64(2): float64(math.Pi)},
+			obj: map[interface{}]interface{}{"one": "Hello", int64(2): float64(math.Pi)},
 		},
 		{
 			dec: map[interface{}]interface{}{"one": "Hello", "two": map[interface{}]interface{}{"three": "Test", "four": "Embedded"}},
